Move auth logic out of component and into services
